Update env_var.md to correct DBT_PASSWORD to DBT_ENV_SECRET_<PASSWORD>
In the following:

"import Envvarsecrets from '/snippets/_env-var-secrets.md'; 

<Envvarsecrets />

If the `DBT_USER` and `DBT_PASSWORD`..."

If the `DBT_ENV_SECRET_` prefix is not used, I believe the user's password will be shown on compilation or in logs. I believe the best practice is to use the `DBT_ENV_SECRET_` prefix for secrets, passwords, etc.

The '/snippets/_env-var-secrets.md' will also need to be modified if this change is to be adopted.